requirements.txt
fastapi
uvicorn[standard]
python-multipart
redis
rq
python-dotenv
